feat: make photo selection into a composable
